Updated the following AWS CloudFormation resources:
 - AWS::EC2::TrafficMirrorFilterRule
 - AWS::Cognito::UserPoolDomain
 - AWS::SageMaker::Workteam.NotificationConfiguration
 - AWS::Cognito::UserPoolRiskConfigurationAttachment.NotifyConfigurationType
 - AWS::Config::OrganizationConfigRule.OrganizationCustomRuleMetadata
 - AWS::ECS::TaskDefinition.ContainerDefinition
 - AWS::Config::RemediationConfiguration.ExecutionControls
 - AWS::ApplicationAutoScaling::ScalableTarget.SuspendedState
 - AWS::Cognito::UserPoolIdentityProvider
 - AWS::SageMaker::CodeRepository
 - AWS::RDS::DBCluster
 - AWS::ECS::TaskDefinition.Tmpfs
 - AWS::Glue::Trigger.Condition
 - AWS::Events::Rule.EcsParameters
 - AWS::Lambda::EventSourceMapping
 - AWS::EC2::TrafficMirrorSession
 - AWS::DynamoDB::Table.SSESpecification
 - AWS::AppMesh::Route.Duration
 - AWS::Glue::Job.NotificationProperty
 - AWS::Pinpoint::App
 - AWS::ECS::TaskDefinition
 - AWS::EC2::VPCEndpointService
 - AWS::Pinpoint::Segment
 - AWS::SSM::MaintenanceWindowTarget.Targets
 - AWS::SageMaker::Workteam
 - AWS::Config::OrganizationConfigRule
 - AWS::Cognito::UserPoolUICustomizationAttachment
 - AWS::EC2::Instance.Ebs
 - AWS::EC2::TrafficMirrorFilterRule.TrafficMirrorPortRangeRequest
 - AWS::Cognito::UserPoolResourceServer.ResourceServerScopeType
 - AWS::Budgets::Budget.BudgetData
 - AWS::Cognito::UserPoolRiskConfigurationAttachment.RiskExceptionConfigurationType
 - AWS::Cognito::UserPoolResourceServer
 - AWS::DocDB::DBCluster
 - AWS::Config::RemediationConfiguration
 - AWS::SSM::MaintenanceWindowTarget
 - AWS::Events::Rule.NetworkConfiguration
 - AWS::Cognito::UserPoolRiskConfigurationAttachment.AccountTakeoverActionType
 - AWS::Cognito::UserPoolRiskConfigurationAttachment.CompromisedCredentialsActionsType
 - AWS::Glue::Trigger.NotificationProperty
 - AWS::Cognito::UserPool
 - AWS::Glue::Job
 - AWS::Glue::Trigger.Action
 - AWS::RDS::DBCluster.DBClusterRole
 - AWS::AppMesh::Route.HeaderMatchMethod
 - AWS::AppMesh::Route.MatchRange
 - AWS::ElastiCache::ReplicationGroup
 - AWS::Cognito::UserPoolRiskConfigurationAttachment
 - AWS::Cognito::UserPoolRiskConfigurationAttachment.AccountTakeoverRiskConfigurationType
 - AWS::Cognito::UserPoolRiskConfigurationAttachment.NotifyEmailType
 - AWS::AppSync::GraphQLApi.LogConfig
 - AWS::QLDB::Ledger
 - AWS::EC2::TrafficMirrorFilter
 - AWS::Pinpoint::Campaign
 - AWS::AppMesh::Route.HttpRouteHeader
 - AWS::Config::OrganizationConfigRule.OrganizationManagedRuleMetadata
 - AWS::Cognito::UserPool.PasswordPolicy
 - AWS::SageMaker::Workteam.MemberDefinition
 - AWS::Neptune::DBCluster
 - AWS::Cognito::UserPoolRiskConfigurationAttachment.AccountTakeoverActionsType
 - AWS::ECS::TaskDefinition.SystemControl
 - AWS::Events::Rule.AwsVpcConfiguration
 - AWS::RDS::DBInstance
 - Tag
 - AWS::RDS::DBInstance.DBInstanceRole
 - AWS::Cognito::UserPoolRiskConfigurationAttachment.CompromisedCredentialsRiskConfigurationType
 - AWS::Cognito::UserPoolClient
 - AWS::SageMaker::CodeRepository.GitConfig
 - AWS::Cognito::UserPoolDomain.CustomDomainConfigType
 - AWS::Elasticsearch::Domain.ZoneAwarenessConfig
 - AWS::Glue::DevEndpoint
 - AWS::AppMesh::Route.HttpRetryPolicy
 - AWS::SageMaker::Workteam.CognitoMemberDefinition
 - AWS::DMS::ReplicationTask
 - AWS::EC2::Instance
 - AWS::EC2::TrafficMirrorTarget
 - AWS::Events::EventBusPolicy
 - AWS::ECS::TaskDefinition.LogConfiguration
 - AWS::AppMesh::Route.RouteSpec
 - AWS::Cognito::UserPool.AdminCreateUserConfig
 - AWS::AppMesh::Route.HttpRouteMatch
 - AWS::CloudWatch::Alarm
 - AWS::IAM::Role
 - AWS::Glue::Trigger
 - AWS::Elasticsearch::Domain.ElasticsearchClusterConfig
 - AWS::ApplicationAutoScaling::ScalableTarget
 - AWS::EC2::ClientVpnEndpoint.TagSpecification
 - AWS::EC2::Instance.CpuOptions
 - AWS::Glue::Workflow
 - AWS::Config::RemediationConfiguration.SsmControls
 - AWS::AppMesh::Route.HttpRoute

+++ cloudformation/resources/aws-cognito-userpool_passwordpolicy.go
@@ -34,7 +34,7 @@ type AWSCognitoUserPool_PasswordPolicy struct {
 	// TemporaryPasswordValidityDays AWS CloudFormation Property
 	// Required: false
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-cognito-userpool-passwordpolicy.html#cfn-cognito-userpool-passwordpolicy-temporarypasswordvaliditydays
-	TemporaryPasswordValidityDays float64 `json:"TemporaryPasswordValidityDays,omitempty"`
+	TemporaryPasswordValidityDays int `json:"TemporaryPasswordValidityDays,omitempty"`
 
 	// _deletionPolicy represents a CloudFormation DeletionPolicy
 	_deletionPolicy policies.DeletionPolicy
